# internal/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIEnvelope:
    properties:
      data:
        type: object
      message:
        example: ok
        type: string
      success:
        example: true
        type: boolean
    type: object
  api.APIErrorEnvelope:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: true
        type: boolean
      message:
        example: invalid request
        type: string
    type: object
...
  api.OkData:
    properties:
      ok:
        example: true
        type: boolean
    type: object
  api.OkResponse:
    properties:
      data:
        $ref: '#/definitions/api.OkData'
      success:
        example: true
        type: boolean
    type: object
...
  api.SingboxControlRequest:
    properties:
      action:
        enum:
        - start
        - stop
        - restart
        example: start
        type: string
    type: object
...
  api.SingboxStatusData:
    properties:
      currentVersion:
        example: 1.13.11
        type: string
      features:
        example:
        - with_quic
        items:
          type: string
        type: array
      installed:
        example: true
        type: boolean
      pid:
        example: 12345
        type: integer
      proxyComponent:
        example: true
        type: boolean
      requiredVersion:
        example: 1.13.11
        type: string
      running:
        example: true
        type: boolean
      tunnelCount:
        example: 2
        type: integer
      updateAvailable:
        example: false
        type: boolean
      version:
        example: 1.9.3
        type: string
    type: object
  api.SingboxStatusResponse:
    properties:
      data:
        $ref: '#/definitions/api.SingboxStatusData'
      success:
        example: true
        type: boolean
    type: object
...
  api.SingboxTunnelDTO:
    properties:
      connectivity:
        $ref: '#/definitions/api.SingboxTunnelConnectivity'
      fingerprint:
        example: chrome
        type: string
      listenPort:
        example: 7891
        type: integer
      port:
        example: 443
        type: integer
      protocol:
        example: vless
        type: string
      proxyInterface:
        example: br0
        type: string
      running:
        example: true
        type: boolean
      security:
        example: reality
        type: string
      server:
        example: proxy.example.com
        type: string
      sni:
        example: cdn.example.com
        type: string
      tag:
        example: proxy-01
        type: string
      transport:
        example: tcp
        type: string
    type: object
  api.SingboxTunnelsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.SingboxTunnelDTO'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /singbox/control:
    post:
      consumes:
      - application/json
      description: Starts, stops, or restarts the sing-box engine. Returns the fresh
        status snapshot. Mirrors /system/hydraroute-control.
      parameters:
      - description: Action to perform
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SingboxControlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SingboxStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Control sing-box process
      tags:
      - singbox
  /singbox/install:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Install sing-box
      tags:
      - singbox
...
  /singbox/status:
    get:
      description: Available when sing-box integration is enabled in the build.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SingboxStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Sing-box status
      tags:
      - singbox
...
  /singbox/tunnels:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Delete sing-box tunnel
      tags:
      - singbox
    get:
      parameters:
      - description: When set, returns single tunnel
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SingboxTunnelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: List or get sing-box tunnel(s)
      tags:
      - singbox
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Add sing-box tunnel(s)
      tags:
      - singbox
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Update sing-box tunnel
      tags:
      - singbox
  /singbox/tunnels/delay-check:
    post:
      parameters:
      - description: Tunnel tag
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Sing-box delay check
      tags:
      - singbox
  /singbox/tunnels/test/connectivity:
    get:
      description: Tests connectivity through a sing-box tunnel. Provide either `tag`
        (resolved to tunnel kernel interface) or `iface` (direct kernel interface
        override, useful for subscription tests).
      parameters:
      - description: Tunnel tag (required when iface is not set)
        in: query
        name: tag
        type: string
      - description: Kernel interface override (e.g. t2s12)
        in: query
        name: iface
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Sing-box tunnel connectivity test
      tags:
      - singbox
  /singbox/tunnels/test/ip:
    get:
      description: Resolves current external IP through a sing-box tunnel. Provide
        either `tag` (resolved to tunnel kernel interface) or `iface` (direct kernel
        interface override, useful for subscription tests). Optional `service` overrides
        IP-check endpoint.
      parameters:
      - description: Tunnel tag (required when iface is not set)
        in: query
        name: tag
        type: string
      - description: Kernel interface override (e.g. t2s12)
        in: query
        name: iface
        type: string
      - description: Custom IP-check service URL
        in: query
        name: service
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Sing-box tunnel IP test
      tags:
      - singbox
  /singbox/tunnels/test/speed/stream:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Sing-box tunnel speed test stream
      tags:
      - singbox
  /singbox/update:
    post:
      description: Replaces the currently-installed managed sing-box with the version
        this awg-manager build is pinned to. No-op when versions match.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OkResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Update managed sing-box binary
      tags:
      - singbox
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"